+++ src/middleware/next.ts
@@ -69,10 +69,12 @@ export function createNextMiddleware<T extends SignalAttributes>(
   tracer: Tracer,
   schemaVersion: string,
   options: MiddlewareOptions<T>,
+  generateRequestId: () => string,
 ) {
   const handleRequest = createRequestHandler<T>(store, tracer, {
     schemaVersion,
     options,
+    generateRequestId,
   })
 
   return async function nextMiddleware(